Read Me First {pmclust}	R Documentation

Read Me First Function

Description

This function print the annotations of all variables used in this package.

*.dmat's are ddmatrix versions.

Usage

  readme()

  readme.dmat()

Details

This package is optimized in the way by pre-specifying several global variables in .pmclustEnv. These variables will be overwrote by EM algorithms. Users should use these names to access the results and utilize them with cautions.

readme.dmat is a ddmatrix version of readme.

Value

A readme message will print on screen by default and explain the global variables used in this package, including:

CHECK 	 convergent checking

CLASS.spmd 	 true id of each data, a vector of length N.spmd and has values from 1 to K

COMM.RANK 	 rank of current processor, obtained from comm.rank of pbdMPI

COMM.SIZE 	 total processors in MPI world, obtained from comm.size of pbdMPI

CONTROL 	 controls for EM iterations

PARAM 	 set or parameters

SAVE.param 	 (debug only) save parameters for every iterations

SAVE.iter 	 (debug only) save computing time for every iterations

U.spmd 	 temporary storage for density

W.spmd 	 temporary storage for eta * density

W.spmd.rowSums 	 temporary storage for rowSums of W.spmd

X.spmd 	 generated data set with dimension with dimension N.spmd * p

Z.colSums 	 temporary storage for rowSums of Z.spmd

Z.spmd 	 posterior probabilities

p.times.logtwopi 	 p * log(2 * pi)

Each variable may contain several elements if it is a list, some variables are used for temporary storages in order to optimize computing, and some variables are used for constant variables. These variables may be restricted, and only generated by the function set.global.

One can access these variables via the global environment .pmclustEnv such as .pmclustEnv$CONTROL.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

set.global, set.global.dmat.

Examples

## Not run: 
readme()
readme.dmat()

## End(Not run)

[Package pmclust version 0.1-7 Index]