Set Global Variables {pmclust}	R Documentation

Set Global Variables According to the global matrix X.gbd (X.spmd) or X.dmat

Description

This function will set several sets of variables globally in the environment .pmclustEnv according to the global matrix X.gbd/X.spmd.

set.global.dmat is a ddmatrix version of set.global.gbd.

Usage

  set.global.gbd(K = 2, X.gbd = NULL, PARAM = NULL,
      algorithm = c("em", "aecm", "apecm", "apecma", "kmeans"),
      RndEM.iter = 10)
  set.global.dmat(K = 2, X.dmat = NULL, PARAM = NULL,
      algorithm = c("em.dmat", "kmeans.dmat"),
      RndEM.iter = 10)

  set.global(K = 2, X.spmd = NULL, PARAM = NULL,
      algorithm = c("em", "aecm", "apecm", "apecma", "kmeans"),
      RndEM.iter = 10)

Arguments

K
an original set of parameters generated by set.global.

X.gbd
an input GBD matrix.

X.dmat
an input ddmatrix.

X.spmd
an input SPMD matrix.

PARAM
an original set of parameters generated by set.global.

algorithm
an original set of parameters generated by set.global.

RndEM.iter
number of RndEM iterations.

Details

WARNING: A global variable named X.gbd/X.spmd should be set before calling set.global where X.gbd/X.spmd is a matrix containing data with dimension N.spmd * p. i.e. N.spmd observations and p variables.

WARNING: A global variable named X.dmat should be set before calling set.global.dmat where X.dmat is a ddmatrix (in block-cyclic format) containing data with dimension N * p.

X.gbd/X.spmd and X.dmat are supposed to exist in .GlobalEnv. If not, they should be as an input object and will be copied into .pmclustEnv which is less efficient.

Value

A new set of PARAM will be returned and several global variables will be set according to the data X.gbd/X.spmd.

Sets of global variables are store in the default environment .pmclustEnv.

Use readme to see all global variables set by this function.

Use readme.dmat to see all global variables of ddmatrix version set by this function.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

em.step, aecm.step, apecm.step,
apecma.step.

Examples

## Not run: 
# Examples can be found in the help pages of em.step(),
# aecm.step(), apecm.step(), apecma.step(), and kmeans.step().

# Examples for ddmatrix version can be found in the help pages of
# kmeans.step.dmat().

## End(Not run)

[Package pmclust version 0.1-7 Index]