Set of PARAM {pmclust}	R Documentation

A Set of Parameters in Model-Based Clustering.

Description

This set of parameters are used in initialization, EM iterations, and final convergent results. All share the same structure in a list variable.

Format

A list variable contains several parameters for computing.

Details

The elements of PARAM or PARAM.org are

N 	 number of observations

p 	 dimension of each observation, total number of variables

K 	 number of clusters

ETA 	 mixing proportion

log.ETA 	 log of mixing proportion

MU 	 centers, dim = \code{p*K}

SIGMA 	 dispersions, a list containing K elements, each element is a matrix, dim = \code{p*p}

U 	 Choleski of SIGMA, the same size of SIGMA

U.check 	 checks of each elements of U, length K

logL 	 log likelihood

min.N.CLASS 	 minimum number of elements in a cluster (restrictions)

The model parameters are ETA, MU, and SIGMA, while log.ETA, U, U.check, and min.N.CLASS are only used in computing.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

set.global.

Examples

## Not run: 
# Use set.global() to generate one of this.
# X.spmd should be pre-specified before calling set.global().

## End(Not run)

[Package pmclust version 0.1-7 Index]