One M-Step {pmclust}	R Documentation

Compute One M-Step Based on Current Posterior Probabilities

Description

This function will perform one M-step based on current posterior probabilities. This is a core function of em.onestep.

m.step.dmat is a ddmatrix version of m.step.

Usage

  m.step(PARAM)

  m.step.dmat(PARAM)

Arguments

PARAM
a set of parameters.

Details

This function will base on the current posterior probabilities Z.spmd to estimate the parameters PARAM mainly including mixing proportions ETA, centers of clusters MU, and dispersions of clusters SIGMA.

Value

Returning a new PARAM which maximizes the complete data log likelihood for the current iteration.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

set.global, em.onestep, e.step.

Examples

## Not run: 
# This is a core function for em.onestep()
# see the source code for details.

## End(Not run)

[Package pmclust version 0.1-7 Index]