as functions {pmclust}	R Documentation

Convert between X.gbd (X.spmd) and X.dmat

Description

This function will convert a matrix of format from GBD row-major to ddmatrix vice versa.

Usage

  as.dmat(X.spmd, bldim = .pbd_env$BLDIM, ICTXT = .pbd_env$ICTXT,
          comm = .pbd_env$SPMD.CT$comm)

  as.gbd(X.dmat, comm = .pbd_env$SPMD.CT$comm)
  as.spmd(X.dmat, comm = .pbd_env$SPMD.CT$comm)

Arguments

X.spmd
an input dataset of format GBD/SPMD row-major to be converted.

X.dmat
an input dataset of format ddmatrix to be converted.

bldim
block dimension, see pbdBASE and pbdDMAT for details.

ICTXT
block context, see pbdBASE and pbdDMAT for details.

comm
communicator, see pbdMPI for details.

Details

This function will provide a quick conversion to the input data X.spmd which is a default dataset to be clustered in pmclust. Usually, this matrix is distributed in GBD row-major format. For fully utilizing pbdBASE and pbdDMAT, it need to be converted to ddmatrix format.

Value

A ddmatrix is returned.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

kmeans.step.dmat.

Examples

## Not run: 
# Examples can be found in the help pages of
# kmeans.step.dmat().

## End(Not run)

[Package pmclust version 0.1-7 Index]