Independent logL {pmclust}	R Documentation

Independent Function for Log Likelihood

Description

This function is for debugging only and for checking if the observed data log likelihood is consistent for each EM iteration.

indep.logL.dmat is a ddmatrix version of indep.logL.

Usage

  indep.logL(PARAM)

  indep.logL.dmat(PARAM)

Arguments

PARAM
a set of parameters.

Details

This function will provide an observed data log likelihood based on the current parameter PARAM. This function will take in information from global, but no global variables will be updated by this function.

This function also don't take care the numerical issues, so the return value may be inaccurate sometimes.

Value

An observed data log likelihood will be returned. This value can quickly compare with the log likelihood computed inside em.onestep. Small difference is allowed, but large difference indicates bugs of code or illness of data.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

set.global, em.onestep.

Examples

## Not run: 
# This is a core function for em.estep()
# see the source code for details.
# Reset .pmclustEnv$CONTROL$debug to turn on this function
# automatically for each EM iteration.

## End(Not run)

[Package pmclust version 0.1-7 Index]