mb.print {pmclust}	R Documentation

Print Results of Model-Based Clustering

Description

This function will print summarized messages for model-based clustering.

Usage

  mb.print(PARAM, CHECK)

Arguments

PARAM
a set of convergent parameters to be printed.

CHECK
a set of checking parameters to be printed.

Details

This function will provide a quick summary from the PARAM and CHECK typically the output of clusterings when algorithms stop. The COMM.RANK 0 will be used to take care the printing.

Value

Summarized messages will print/cat on screen by default.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

em.step, aecm.step, apecm.step,
apecma.step.

Examples

## Not run: 
# Examples can be found in the help pages of em.step(),
# aecm.step(), apecm.step(), and apecma.step().

## End(Not run)

[Package pmclust version 0.1-7 Index]