Update Class of EM or Kmenas Results {pmclust}	R Documentation

Update CLASS.spmd Based on the Final Iteration

Description

Update CLASS.spmd based on the final iteration of EM-like algorithms.

*.dmat's are ddmatrix versions.

Usage

  em.update.class()
  kmeans.update.class()

  em.update.class.dmat()
  kmeans.update.class.dmat()

Details

This function takes Z.spmd from the global environment .pmclustEnv and update CLASS.spmd, and provides the identification of groups for all data.

Value

CLASS.spmd will be updated.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

em.step, aecm.step, apecm.step,
apecma.step, kmeans.step.
kmeans.step.dmat.

Examples

## Not run: 
# Examples can be found in the help pages of em.step(),
# aecm.step(), apecm.step(), apecma.step(), and kmeans.step().

# Examples for ddmatrix version can be found in the help pages of
# kmeans.step.dmat().

## End(Not run)

[Package pmclust version 0.1-7 Index]