print.object {pmclust}	R Documentation

Functions for Printing or Summarizing Objects According to Classes

Description

Several classes are declared in pmclust, and these are functions to print and summary objects.

Usage

## S3 method for class 'pmclust'
print(x, ...)
## S3 method for class 'pkmeans'
print(x, ...)

Arguments

x
an object with the class attributes.

...
other possible options.

Details

These are useful functions for summarizing.

Value

The results will cat or print on the STDOUT by default.

Author(s)

Wei-Chen Chen wccsnow@gmail.com and George Ostrouchov.

References

High Performance Statistical Computing (HPSC) Website: http://thirteen-01.stat.iastate.edu/snoweye/hpsc/

Programming with Big Data in R Website: http://r-pbd.org/

See Also

pmclust, pkmeans.

Examples

## Not run: 
library(pmclust, quiet = TRUE)

# Functions applied by directly type the names of objects.

## End(Not run)

[Package pmclust version 0.1-7 Index]